Separating out integrators; Runtime err fmt
