+++ include/CMakeLists.txt
@@ -26,6 +26,12 @@ set(DDG_HEADERS
 
     "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/system.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/mesh_mutator.h"
-    "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/integrator.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/trajfile.h"
+    
+    "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/integrator/integrator.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/integrator/forward_euler.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/integrator/conjugate_gradient.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/integrator/bfgs.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/integrator/stormer_verlet.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/integrator/velocity_verlet.h"
     PARENT_SCOPE)
