+++ include/mem3dg/mem3dg
@@ -22,5 +22,12 @@
 
 
 #include "solver/system.h"
-#include "solver/integrator.h"
+#include "solver/mesh_mutator.h"
 #include "solver/trajfile.h"
+
+#include "solver/integrator/integrator.h"
+#include "solver/integrator/velocity_verlet.h"
+#include "solver/integrator/stormer_verlet.h"
+#include "solver/integrator/forward_euler.h"
+#include "solver/integrator/conjugate_gradient.h"
+#include "solver/integrator/bfgs.h"
