+++ src/CMakeLists.txt
@@ -13,20 +13,21 @@
 #
 
 set(DDG_SOURCES
-    "${CMAKE_CURRENT_SOURCE_DIR}/pressure.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/energy.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/regularization.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/init.cpp"
-    
-    "${CMAKE_CURRENT_SOURCE_DIR}/stormer_verlet.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/BFGS.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/velocity_verlet.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/euler.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/conjugate_gradient.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/integrator.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/trajfile.cpp"
-
-    "${CMAKE_CURRENT_SOURCE_DIR}/visualization.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/mesh.cpp"
+
+    "${CMAKE_CURRENT_SOURCE_DIR}/solver/pressure.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/solver/energy.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/solver/regularization.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/solver/trajfile.cpp"
+
+    "${CMAKE_CURRENT_SOURCE_DIR}/solver/integrator/integrator.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/solver/integrator/stormer_verlet.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/solver/integrator/BFGS.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/solver/integrator/velocity_verlet.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/solver/integrator/forward_euler.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/solver/integrator/conjugate_gradient.cpp"
+    
+    "${CMAKE_CURRENT_SOURCE_DIR}/init.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/visualization.cpp"
     PARENT_SCOPE
 )
